+++ src/cmd/hostnames.ts
@@ -1,16 +1,18 @@
-import { root, getAppName } from './root'
-import { API } from './api'
+import { root, getAppName, CommonOptions, addCommonOptions } from './root'
+import { apiClient } from './api'
 import { processResponse } from '../utils/cli'
+import { Command } from 'commandpost';
 
-interface HostnamesOptions { }
+interface HostnamesOptions extends CommonOptions { }
 interface HostnamesArgs { }
 
 const hostnames = root
   .subCommand<HostnamesOptions, HostnamesArgs>("hostnames")
   .description("Manage Fly app's hostnames.")
-  .action(async (opts, args, rest) => {
+  .action(async function (this: Command<HostnamesOptions, HostnamesArgs>, opts, args, rest) {
+    const API = apiClient(this)
     try {
-      const res = await API.get(`/api/v1/apps/${getAppName()}/hostnames`)
+      const res = await API.get(`/api/v1/apps/${getAppName(this, { env: ['production'] })}/hostnames`)
       processResponse(res, (res: any) => {
         if (!res.data.data || res.data.data.length === 0)
           return console.log("No hostnames configured, use `fly hostnames add` to add one.")
@@ -31,12 +33,13 @@ interface HostnamesAddArgs {
   hostname: string
 }
 
-hostnames
+const hostnamesAdd = hostnames
   .subCommand<HostnamesAddOptions, HostnamesAddArgs>("add <hostname>")
   .description("Add a hostname to your fly app.")
-  .action(async (opts, args, rest) => {
+  .action(async function (this: Command<HostnamesOptions, HostnamesArgs>, opts, args, rest) {
+    const API = apiClient(this)
     try {
-      const res = await API.post(`/api/v1/apps/${getAppName()}/hostnames`, { data: { attributes: { hostname: args.hostname } } })
+      const res = await API.post(`/api/v1/apps/${getAppName(this, { env: ['production'] })}/hostnames`, { data: { attributes: { hostname: args.hostname } } })
       processResponse(res, (res: any) => {
         console.log(`Successfully added hostname ${args.hostname}`)
       })
@@ -48,4 +51,5 @@ hostnames
     }
   })
 
-
+addCommonOptions(hostnames)
+addCommonOptions(hostnamesAdd)