+++ test/bridge/logger.spec.ts
@@ -37,15 +37,15 @@ describe('logger', function () {
       const msg2Str: string = msgs[1].toString()
       expect(msg2Str).to.include("debug: debug test")
       expect(msg2Str).to.include("method=GET")
-      expect(msg2Str).to.include("foo=bar")
-      expect(msg2Str).to.include("hello=notworld")
+      expect(msg2Str).to.include("foo: 'bar'")
+      expect(msg2Str).to.include("hello: 'notworld'")
 
       const msg3Str: string = msgs[2].toString()
       expect(msg3Str).to.include("error: Error: oh no")
       // sourcemap
       expect(msg3Str).to.include("v8env/events.js")
       expect(msg2Str).to.include("method=GET")
-      expect(msg2Str).to.include("foo=bar")
+      expect(msg2Str).to.include("foo: 'bar'")
     })
 
   })